配置linux环境下的环境部署
Linux环境：
Linux localhost.localdomain 2.6.18-238.el5 #1 SMP Sun Dec 19 14:24:47 EST 2010 i686 i686 i386 GNU/Linux
确保你已经安装了gcc和gawk。
步骤
1.到www.apuebook.com下载源码
2.tar解包，cd apue.2e
3.vi Make.defines.linux   修改变量WKDIR，指向你的apue源码的位置，我的是/home/huangz/code/apue.2e，所以
  WKDIR=/home/huangz/code/apue.2e
4.vi include/apue.h  增加一个常量ARG_MAX，这是threadctl/getenv1.c和threadctl/getenv3.c要用到的；4096这个值是参考里给的，如果有问题，自己修改吧。
  #define ARG_MAX 4096
5.vi threadctl/getenv1.c   增加
  #include "apue.h"
6.vi threadctl/getenv3.c   增加
  #include "apue.h"
7.vi threads/badexit2.c   修改第31行，将pthread_self()的返回值转换为int类型。
  printf("thread 2: ID is %d\n", (int)pthread_self());
8.vi std/linux.mk   将两个nawk改为gawk
9.make
10.sudo cp include/apue.h /usr/include
   sudo cp lib/libapue.a /usr/lib
好了，测试一下，记得要用-lapue命令让编译器链接apue库
gcc main.c -lapue

+++ threads/badexit2.c
@@ -28,7 +28,7 @@ thr_fn1(void *arg)
 void *
 thr_fn2(void *arg)
 {
-	printf("thread 2: ID is %d\n", pthread_self());
+	printf("thread 2: ID is %d\n",(int)pthread_self());
 	pthread_exit((void *)0);
 }
